requirements.txt
pillow
matplotlib
numpy
tensorflow
keras
h5py
shap